requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.7
python-dotenv
fastapi==0.103.1
uvicorn==0.23.2
pendulum
python-multipart==0.0.6
pyjwt==2.1.0





